set default `base-z` to 3000.
originally we use `modal` if zmgr is provided.
this scope the zmgr into `modal`, which may have issues if we want to
manage z-index along with loaders or other elements.
additionally, user will have to remember to config their manually
created ldcover with `modal` `base-z`, otherwise the z-index won't
align.

this really causes some issue so instead of hard code a default `modal`,
we use 3000 as default value for user to overwrite.